requirements.txt
streamlit
scikit-image
scikit-learn
opencv-python-headless
pillow
joblib
numpy
pandas
